backend/requirements.txt
fastapi==0.115.6
starlette==0.41.2
uvicorn==0.23.2
sqlalchemy==2.0.48
pydantic==2.12.5
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.22
pandas==2.1.3
openpyxl==3.1.2
alembic==1.12.1
python-dotenv==1.0.0
